+++ src/infrastructure/three/grids/IGridStyle.ts
@@ -52,6 +52,7 @@ export interface IGridOptions {
 
 export enum GridType {
   SHADOW_FLOOR = 'shadow_floor',
+  CONCRETE_DISC = 'concrete_disc',
   SQUARE_WIRE = 'square_wire',
   HEXAGONAL_WIRE = 'hexagonal_wire',
   HEXAGONAL_GLASS = 'hexagonal_glass',